Fix SEMI_SEMI_AND bug inside $() command substitutions.
Consider the following four uses of the case statement.

1. case 1 in 1) echo Y;; esac
2. case 1 in 1) echo Y;;& esac
3. echo $(case 1 in 1) echo Y;; esac)
4. echo $(case 1 in 1) echo Y;;& esac)

Items 1, 2, and 3 work as expected, while item 4 drops
to PS2 and hangs.

From man(1) bash:

> Using ;;& in place of ;; causes the shell to test the next
> pattern list in the statement, if any, and execute any
> associated list on a successful match.

The phrase "if any" suggests that the success of item 2
is the intended behavior, whereas the failure of item 4
is not intentional.

This commit modifies parse.y so that item 4 works as
intended.